feat(CameraRig): support default SteamVR prefab
It is now possible to use the default SteamVR `[CameraRig]` prefab and
the relevant dependencies on the prefab should now be set up at runtime
via the relevant scripts.

An example scene has been added to demonstrate the prefab support.

It is likely in the future that the custom `[CameraRig]` prefab
provided by the toolkit will be deprecated and removed as it should not
be required. At this point, the example scenes will be updated to use
the default SteamVR `[CameraRig]` and
`0034_CameraRig_SteamVRPrefabSupport` will be removed.